Adding ExerciseView and modify navigation using Navigation Stack
